mutant_api/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
click==8.1.7
colorama==0.4.6
dill==0.3.9
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==5.13.2
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
mysqlclient==2.2.5
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.32.0
